Binary Trees
We extend the concept of linked data structures to structure containing nodes with more than one self-referenced field. A binary tree is made of nodes, where each node contains a "left" reference, a "right" reference, and a data element. The topmost node in the tree is called the root.

Every node (excluding a root) in a tree is connected by a directed edge from exactly one other node. This node is called a parent. On the other hand, each node can be connected to arbitrary number of nodes, called children. Nodes with no children are called leaves, or external nodes. Nodes which are not leaves are called internal nodes. Nodes with the same parent are called siblings.
	   

More tree terminology:

    The depth of a node is the number of edges from the root to the node.
    The height of a node is the number of edges from the node to the deepest leaf.
    The height of a tree is a height of the root.
    A full binary tree.is a binary tree in which each node has exactly zero or two children.
    A complete binary tree is a binary tree, which is completely filled, with the possible exception of the bottom level, which is filled from left to right.